refactor: consolidate query and accessors into unified query module
- Merge accessors/index.ts into query/index.ts
- Remove separate accessors/ directory
- Update all exports in index.ts to use query/
- Update tsup.config.ts to remove accessors entry point
- Fix test imports to use query/ instead of accessors/

This simplifies the API structure:
- query/ - all read operations (search, get, iterate, aggregate)
- operations/ - all write operations (add, delete, modify)
