+++ tsup.config.ts
@@ -3,7 +3,6 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
   entry: {
     index: 'src/index.ts',
-    'accessors/index': 'src/accessors/index.ts',
     'operations/index': 'src/operations/index.ts',
     'query/index': 'src/query/index.ts',
   },
